pyproject.toml
[project]
name = "socialsyncv1"
version = "0.1.0"
description = "Multi-platform Social Media Scheduler"
requires-python = "==3.9.6"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.35",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "cloudinary>=1.40.0",
    "python-multipart>=0.0.9",
    "cryptography>=43.0.0",
    "PyJWT>=2.9.0",
    "pillow>=10.0.0",  # for image processing
]

[tool.uv]
dev-dependencies = [
    "black>=24.0",
    "ruff>=0.5.0",
    "pytest>=8.0",
    "httpx>=0.27.0",
]
